requirements/base.txt
Django==4.2.0
djangorestframework==3.14.0
django-environ==0.11.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
django-cors-headers==4.2.0
whitenoise==6.4.0
Pillow==9.5.0
psycopg2-binary==2.9.10